requirements.txt
pandas
requests
pydantic
pyyaml
openpyxl
sqlalchemy
rapidfuzz
fastapi
uvicorn
pytest
httpx
